Factor out restart logic into language runtime service
